adds --test-user flag for DockerDeploy
Some Docker containers are built with the intent that the user running
the process within the container will *not* be root. While this is a
good security practice, it can cause problems when tests require
additional packages to be installed. The `--test-user` flag allows tests
to be executed as a privileged users for cases where a Docker
container's directory permissions are lacking.