ensures test distribution is consistent each run
When sorting tests exclusively by time the possibility of missing or
duplicating tests was introduced (when tests were marked with the same
estimated duration). This commit uses the test file name as a fallback
sort value, ensuring that test are distributed in the same order (and to
the same nodes) on each run.