adds waiters to avoid race conditions with ECS API
From AWS docs:

The Amazon ECS API follows an eventual consistency model, due to the
distributed nature of the system supporting the API. This means that the
result of an API command you run that affects your Amazon ECS resources
might not be immediately visible to all subsequent commands you run. You
should keep this in mind when you carry out an API command that
immediately follows a previous API command.