BT-480 - migration workflow (#1)
From AWS docs:

```
The Amazon ECS API follows an eventual consistency model, due to the
distributed nature of the system supporting the API. This means that the
result of an API command you run that affects your Amazon ECS resources
might not be immediately visible to all subsequent commands you run. You
should keep this in mind when you carry out an API command that
immediately follows a previous API command.
```

Because of this a `wait` function was introduced to abstract some of the retry/delay patterns needed for operations.

This commit moves build/push/update_task/migrate/update_service to
individual methods to allow more precise selection of operations to the
end user.

The main purpose of this commit is to introduce a "migrate" workflow for running migration tasks before updating service definitions. See the README for examples or `nyprsetuptools DockerDeploy --help`.